+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Bugfix: The `gather-logs` command will no longer send any logs through `gRPC`.
 
+- Change: Telepresence agents watcher will now only watch namespaces that the user has accessed since the last `connect`.
+
 ### 2.5.5 (April 8, 2022)
 
 - Change: The traffic-manager now requires permissions to read pods across namespaces even if installed with limited permissions